requirements.txt
langchain
langchain_experimental
langchain-google-genai
python-dotenv
streamlit
tiktoken
faiss-cpu
protobuf
mysql-connector-python
pymysql
sentence-transformers
chromadb






